+++ packages/react-hooks/src/use-dataviews-layers/index.ts
@@ -1,1 +1,1 @@
-export { default, getDataviewsGeneratorConfigs, getGeneratorConfig } from './use-dataviews-layers'
+export { default } from './use-dataviews-layers'